requirements.txt
flask-multipass>=0.4.*
indico>=3.2
python3-saml>=1.10.1,<1.11
flask_sqlalchemy
